comment schema

+++ controllers/admin.js
@@ -9,6 +9,10 @@ exports.getAllPosts = async (req, res, next) => {
         path: "author",
         select: "_id",
       })
+      .populate({
+        path: "comments",
+        select: "description likes dislikes"
+      })
       .sort({ updatedAt: -1, likes: -1, comments: -1, dislikes: 1 });
     res.status(200).json({
       message: "Fetched posts successfully.",
@@ -25,7 +29,11 @@ exports.getAllPosts = async (req, res, next) => {
 exports.getPost = async (req, res, next) => {
   const postId = req.params.postId;
   try {
-    const post = await Post.findById(postId).populate("author");
+    const post = await Post.findById(postId).populate("author")
+    .populate({
+      path: "comments",
+      select: "description likes dislikes"
+    });
     if (!post) {
       const error = new Error("Could not find post.");
       error.statusCode = 404;
@@ -50,6 +58,7 @@ exports.adminDeletePost = async (req, res, next) => {
       throw error;
     }
     if (post.dislikes.totalDislikes > 2) {
+      await Comment.deleteMany({ _id: post.comments });
       await Post.findByIdAndRemove(postId);
       const user = await User.findById(post.author);
       user.posts.pull(postId);
@@ -75,13 +84,16 @@ exports.deleteComment = async (req, res, next) => {
       error.statusCode = 404;
       throw error;
     }
-    const comment = post.comments.find((c) => c.id === commentId);
-    post.comments.pull(comment);
-    await Comment.findByIdAndRemove(commentId);
-    await post.save();
-    res.status(200).json({
-      message: "Comment Deleted Successfully",
-    });
+    const comment = post.comments.find((c) => c.toString() === commentId);
+    if (!comment) {
+      return res.status(404).json({ message: "Comment not found" });
+    }
+      post.comments.pull(commentId);
+      await post.save();
+      await Comment.findByIdAndRemove(commentId);
+      return res.status(200).json({
+        message: "Comment Deleted Successfully",
+      });
   } catch (err) {
     if (!err.statusCode) {
       err.statusCode = 500;
